Final Project for CIS 3260 GAU
 Validating Credit Cards: problem description

Credit card numbers follow certain patterns. A credit card number must have between 13 and 16 digits. It must start with:

4 for Visa cards

5 for Master cards

37 for American Express cards

6 for Discover cards


In 1954, Hans Luhn of IBM proposed an algorithm for validating credit card numbers. The algorithm is useful to determine if a card number is entered correctly or if a credit card is scanned correctly by a scanner. Almost all credit card numbers are generated following this validity check, commonly known as the Luhn check or the Mod 10 check, which can be described as follows (for illustration, consider the card number 4388576018402626):


1. Double every second digit from right to left. If doubling of a digit results in a two-digit number, add up the two digits to get a single-digit number.

2. Now add all single-digit numbers from Step 1.

3. Add all digits in the odd places from right to left in the card number.

4. Sum the results from Step 2 and Step 3.

5. If the result from Step 4 is divisible by 10, the card number is valid; otherwise, it is invalid.


Write a program that prompts the user to enter a credit card number as a long integer. Display whether the number is valid or invalid.